CSE310_w2_2_27947
Shruti and Priyanka are playing a game. Shruti gives two numbers in range of 0-127. Priyanka have to tell that whether first number can be converted to second number by applying left shift operator on the binary representation of first number. Help the priyanka to solve the problem.

Input Format

Two space seperated integer values in the range of 0 to 127.

Constraints

Input will be positive and in range of 0 to 127.

Output Format

Yes if number can be converted and No if cannot be converted.

Sample Input 0

5 20
Sample Output 0

Yes
Sample Input 1

150 30
Sample Output 1

Invalid Input