Week2_26699_1
Sonal built up a software code taking ratings on a product. The product is supposed to be rated between 1 to 5. Create a program displaying an error "101" if the rating is greater than 5, else displaying "thanks".

Input Format

Rating as a number

Constraints

Input can be floating point number or an integer.

Output Format

Error or thanks should be displayed.