CSE310_27947_week5_1
Richa and her daughter Ahaana are playing a game. Richa is going to tell one number and Ahaana need to tell the prime factors of the number. Help Ahaana by completing the code to find prime factor of the number. Write a method which calculate prime factors and print and call the method in main.

Input Format

An integer value

Constraints

N will be lie between 10-50

Output Format

All the prime factors will be printed exectly once with space.