Week8_24906_2
Adwit is playing game of number division with her friend , taking 2 input numbers both are of float type only. He defines rules for game is like that if second number is 0.0 then it will print Infinity and if second number is not 0.0 then it simply print its result and that is also of float type.

Input Format

In first line , enter first number. In second line , enter second number

Constraints

Both the input numbers must be of float type

Output Format

Output depends upon second number, if it is 0.0 then it will print Infinity , but if second number is not zero then it will print result of float division number.

Sample Input 0

8.9
0.0
Sample Output 0

Infinity